runnable module output not deleted if file is not in temp dir

+++ src/ce_main.pas
@@ -676,6 +676,8 @@ begin
     exit;
   //
   fname := fRunProc.Executable;
+  if ExtractFileDir(fname) <> GetTempDir(false) then
+    exit;
   killProcess(fRunProc);
   if fileExists(fname) then
     sysutils.DeleteFile(fname);
